remove short createrow

+++ SWB4/SWBSocialWeb/SWBSocialResources/src/org/semanticwb/social/admin/resources/PieChart.java
@@ -311,12 +311,12 @@ public class PieChart extends GenericResource {
             sheet.addMergedRegion(new CellRangeAddress(0, 0, 0, 13));
 
             // creo una nueva fila
-            Row trow = sheet.createRow((short) 0);
+            Row trow = sheet.createRow(0);
             createTituloCell(wb, trow, 0, CellStyle.ALIGN_CENTER,
                     CellStyle.VERTICAL_CENTER, "Mensajes " + title);
 
             // Creo la cabecera de mi listado en Excel
-            Row row = sheet.createRow((short) 2);
+            Row row = sheet.createRow(2);
 
             // Creo las celdas de mi fila, se puede poner un diseño a la celda
 
@@ -348,7 +348,7 @@ public class PieChart extends GenericResource {
 
                 PostIn postIn = (PostIn) setso.next();
 
-                Row troww = sheet.createRow((short) i);
+                Row troww = sheet.createRow(i);
 
                 if (postIn.getMsg_Text() != null) {
                     if (postIn.getMsg_Text().length() > 2000) {
